dnsdist: Move completion-related code to a separate file
